second session- split to different modules
