+++ training/src/app/app.module.ts
@@ -1,35 +1,11 @@
 import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
-import { HttpClientModule } from "@angular/common/http";
-import { MatDividerModule } from "@angular/material/divider";
-import { MatButtonModule } from "@angular/material/button";
-import { MatInputModule } from "@angular/material/input";
-import { MatOptionModule } from "@angular/material/core";
-import { MatSelectModule } from "@angular/material/select";
-import { MatButtonToggleModule } from "@angular/material/button-toggle";
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ChuckNorrisMainComponent } from "./chuck-norris-main/chuck-norris-main.component";
-import { JokeGeneratorComponent } from './joke-generator/joke-generator.component';
-import { JokeComponent } from './joke/joke.component';
+import { ChuckNorrisModule } from "./modules/chuck-norris/modules/chuck-norris.module";
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    ChuckNorrisMainComponent,
-    JokeGeneratorComponent,
-    JokeComponent,
-  ],
-  imports: [
-    HttpClientModule,
-    BrowserAnimationsModule,
-    MatDividerModule,
-    MatButtonModule,
-    MatButtonToggleModule,
-    MatInputModule,
-    MatOptionModule,
-    MatSelectModule,
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
+	declarations: [AppComponent],
+	imports: [ChuckNorrisModule],
+	providers: [],
+	bootstrap: [AppComponent]
 })
 export class AppModule {}
