+++ training/src/app/modules/chuck-norris/components/chuck-norris-main/chuck-norris-main.component.html
@@ -1,7 +1,7 @@
 <div class="top-bar">
-    <img id="top-bar-fire" src="./assets/fire-img.svg" alt="error">
-    <img id="top-bar-logo" src="./assets/chuck-norris-img.png" alt="error">
-    <img id="top-bar-title" src="./assets/title-img.png" alt="error">
+    <img id="top-bar-fire" src="../../../../../assets/fire-img.svg" alt="error">
+    <img id="top-bar-logo" src="../../../../../assets/chuck-norris-img.png" alt="error">
+    <img id="top-bar-title" src="../../../../../assets/title-img.png" alt="error">
 </div>
 
 <div class="buttons">
