+++ training/src/app/modules/chuck-norris/components/chuck-norris-main/chuck-norris-main.component.scss
@@ -1,4 +1,4 @@
-@import "src/styles";
+@import "../../../../../styles";
 
 .top-bar {
   position: relative;
