+++ training/src/app/modules/chuck-norris/components/chuck-norris-main/chuck-norris-main.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { JokeService } from "./services/joke/joke.service";
-import { Joke } from "../models/joke";
+import { JokeService } from "../../services/joke/joke.service";
+import { Joke } from "../../models/joke.modle";
 
 @Component({
 	selector: 'app-chuck-norris-main',
