+++ training/src/app/modules/chuck-norris/components/joke-generator/joke-generator.component.scss
@@ -1,4 +1,4 @@
-@import "src/styles";
+@import "../../../../../styles";
 
 .buttons-container {
   display: grid;
