+++ training/src/app/modules/chuck-norris/components/joke/joke.component.scss
@@ -1,4 +1,4 @@
-@import "src/styles";
+@import "../../../../../styles";
 
 blockquote {
   border: 2px solid $blue-color;
