+++ training/src/app/modules/chuck-norris/components/joke/joke.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Joke } from "../models/joke";
+import { Joke } from "../../models/joke.modle";
 
 @Component({
 	selector: 'app-joke',
