+++ training/src/app/modules/chuck-norris/models/joke.modle.ts
